requirements.txt
aiosqlite==0.18.0
fastapi==0.88.0
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.38.3
pydantic==1.10.2
databases==0.7.0
python-dateutil==2.8.2
